requirements.txt
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
bson==0.5.10
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.12
Flask==1.1.2
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-Mail==0.9.1
ibm-cloud-sdk-core==3.7.0
ibm-watson==5.1.0
idna==2.10
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
passlib==1.7.4
pendulum==2.1.2
protobuf==3.15.6
py-buzz==1.0.3
PyJWT==1.7.1
pymongo==3.11.3
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytzdata==2020.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
urllib3==1.26.4
websocket-client==0.48.0
Werkzeug==1.0.1
wrapt==1.12.1
